+++ jgclark.DashboardReact/src/react/components/Dialog.jsx
@@ -33,11 +33,11 @@ const Dialog = ({ isOpen, onClose, isTask, details }: Props): React$Node => {
   const { reactSettings, pluginData } = useAppContext()
 
   // 5s hack timer to work around cache not being reliable (only runs for users, not DEVs)  
-  function onDialogClose() {
+  function onDialogClose(xWasClicked: boolean) {
     // Send 'refresh' action to plugin after [5000] ms - this is a bit of a hack
     // to get around the updateCache not being reliable.
     // only forces a refresh if not in logLevel === DEV
-    onClose() // send to parent
+    onClose(xWasClicked) // send to parent
   }
   
   // the child dialogs (Task & Project) will call this function to position the dialog